Make errors terminate script
This changes to use `set -e` mode in the build.bash script.  This makes
any errors more obvious since it will exit the script.

(I had not cloned the repository with `--recurse-submodules` and so
the httpd server folder was empty, but this did not stop the APK from
being created.)