Generalize /testwrite/xxx/yyy.txt command
It turns out that applications at the higher API levels do not have
write access to /data/local/tmp.  In order to make it easier to find
a good place that the app actually can write (SD card? app-specific
data directories?) this changes the /testwrite hack so that it will
use whatever comes after it as the path to try writing a test file to.

This also traps errors and returns them in the http response to make
it easier to see what's going wrong...and also to not take down the
server if the write fails.