AND and OR updates
The AND and OR operators have changed to being LOGIC! only.  They also
now take GROUP! on the right hand side, not BLOCK!...even though the
GROUP! is quoted:

https://forum.rebol.info/t/naming-of-bitwise-or-and-xor/472/14

(Please notice in particular the implications of using GROUP! branches
with IF or EITHER, how non-taken branches are skipped...)

This changes an AND to the GROUP! style, and uses an ELSE instead of
an OR for a case that wasn't expecting a LOGIC! result.