FILE pathing and dirs-need-slashes changes
The old idea of doing file pathing with plain path selection has been
killed off.  There is a new MAKE FILE! and operator form of %% being
worked out:

https://forum.rebol.info/t/can-make-file-dialect-solve-the-madness/1356
https://forum.rebol.info/t/make-file-and-unleashed-time-to-use-it/1398

To make these operators offer some use, a file component can't be used
as a directory unless it ends in `/`.  This helps avoid the implications
of doubling slashes and the confusions related to that.

However Java's getAbsolutePath() does not offer a terminal / on paths,
so this has to add one in the Java.