Proposed direction for (U)PARSE of command line
New features are being explored in UPARSE that are designs that are
expected to be incorporated in PARSE.  Although UPARSE is too slow for
large scale use at this time, getting practical experience on small
pieces of real code will help inform the features.

This demonstrates a new feature of "SET" (now done with SET-WORD!)
that is able to evaluate to the last value in a parse rule.  This is
similar to how DO expressions evaluate to the last value in the rule.
Some rules (such as plain GROUP!) are invisible, but when a GROUP! is
used with an @(...) then the evaluated value is used.

So UPARSE can be used as more of a kind of SWITCH statement.  This
shows that off by using @() expressions instead of repeating the name
of variables to assign.  Being able to put the name outside a block
also assists in writing reusable parse rules.

Another feature is that strings can invoke TRANSCODE.  This makes it
easier to extract an INTEGER! out of a port number.  It's a demo of
how much more useful it is to be able to have a rule produce values
that are synthesized from something besides the input.

An aspect used here is that INTO is arity-2 and able to specify what it
is going to parse into, which can also be a synthesized value:

https://forum.rebol.info/t/new-more-powerful-arity-2-into-in-uparse/1555

It makes use of the @[...] rule to capture across input, as well.