Use UPARSE to process webserver.reb command line
(Note: Although UPARSE is too slow for large scale use at this time,
getting practical experience on small pieces of real code will help
inform the features.)

This takes advantage of UPARSE's ability to not only have parsers
advance through the input, but also synthesize values...with a BLOCK!
rule evaluating to the last synthesized result.  This can include
the result of GROUP!s if they are to be used.

Hence UPARSE can be used as more of a kind of SWITCH statement.  This
shows that off by using expressions instead of repeating the name
of variables to assign.  Being able to put the name outside a block
also assists in writing reusable parse rules.

Another feature is that strings can invoke TRANSCODE.  This makes it
easier to extract an INTEGER! out of a port number.  It's a demo of
how much more useful it is to be able to have a rule produce values
that are synthesized from something besides the input.

An aspect used here is that INTO is arity-2 and able to specify what it
is going to parse into, which can also be a synthesized value:

https://forum.rebol.info/t/new-more-powerful-arity-2-into-in-uparse/1555

A new and very experimental feature of calling functions to process
results of a rule without needing to name variables is also used:

https://forum.rebol.info/t/1608/2