Fix PARSE references
The native PARSE command now returns a BAD-WORD! isotope on success
(in most cases), in order to draw attention to the fact that one cannot
necessarily know that the result of a successful parse will be truthy
or not, now that parses are synthesizing arbitrary values from their
rule blocks.

To work around this for those needing a logic result, there is PARSE?.
It is also possible to use THEN/ELSE, which will reliably work on the
results even if the rules return null, because they will return a
~null~ isotope and hence not be considered "true null" by THEN or ELSE
on success.

Also, the RETURN combinator was removed from UPARSE as it is somewhat
problematic.